+++ src/interfaces/index.tsx
@@ -1,4 +1,4 @@
-export { default } from './Props'
-export { Item } from './Item'
-export { FadeMarker, BorderMarker } from './Marker'
 export { Gradient } from './Gradient'
+export { Item } from './Item'
+export { Marker } from './Marker'
+export { Props } from './Props'